Split service_coverage_panel.rs into subdirectory modules (1296 -> ~400 lines each)
Break the monolithic service_coverage_panel.rs into logical sub-modules:
- categories.rs (240 lines): ServiceCategory, OtherServiceGroup enums and impls
- stats.rs (138 lines): CategoryStats, ServiceTypeStats, computation functions, color helpers
- panel_ui.rs (377 lines): UI systems, keybind, visibility resource, plugin
- tests.rs (333 lines): Category, color, label, overlay, and state tests
- tests_computation.rs (220 lines): Coverage computation and per-type stats tests
- mod.rs (31 lines): Re-exports all pub items for transparent access
